pyproject.toml
[project]
name = "devops-engineer-from-scratch-project-313"
version = "0.2.0"
description = "FastAPI project for Hexlet DevOps"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.121.2",
    "fastapi-cli>=0.0.16",
    "uvicorn>=0.38.0",
    "pytest>=7.0.0",
    "ruff>=0.0.293",
    "httpx>=0.24.0",
    "sentry-sdk[fastapi]>=2.46.0",
    "sqlmodel>=0.0.16",
    "python-dotenv>=1.0.1",
    "psycopg[binary]>=3.2.1",
]


